pyproject.toml
[project]
name = "parse-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncio>=4.0.0",
    "google-genai>=1.57.0",
    "openai>=2.15.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
